requirements/base.txt
# Bleeding edge Django
django==1.9.0

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.6


# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.6.0

# Models
django-model-utils==2.3.1

# Images
Pillow==2.9.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.24.1


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.7

# Redis support
django-redis==4.3.0
redis>=2.10.0



# Your custom requirements go here
#API Setup
djangorestframework==3.3.2
django-filter==0.12.0
Markdown==2.6.5
